requirements.txt
altair==5.3.0
asn1crypto==1.5.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contourpy==1.0.7
cryptography==42.0.8
cycler==0.11.0
distlib==0.3.8
filelock==3.13.1
Flask==3.0.0
fonttools==4.39.3
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
numpy==1.24.3
packaging==23.1
pandas==2.2.2
Pillow==9.5.0
platformdirs==4.1.0
protobuf==5.27.2
pyarrow==16.1.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.19.0
six==1.16.0
smmap==5.0.1
snowflake-connector-python==3.11.0
snowflake-snowpark-python==1.19.0
sortedcontainers==2.4.0
streamlit==1.36.0
tenacity==8.5.0
toml==0.10.2
tomlkit==0.13.0
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
virtualenv==20.25.0
Werkzeug==3.0.1
